+++ bfabric_app_runner/src/bfabric_app_runner/output_registration/register.py
@@ -89,12 +89,11 @@ def _identify_existing_resource_id(
         # TODO maybe it would be more accurate to use relativepath here, however historically it would often start
         #      with `/` which can be confusing.
         resource = client.reader.query_one(
-            "resource",
+            Resource,
             {
                 "name": spec.store_entry_path.name,
                 "workunitid": workunit_definition.registration.workunit_id,
             },
-            expected_type=Resource,
         )
         if resource is not None:
             return resource.id
@@ -168,7 +167,7 @@ def _save_dataset(spec: SaveDatasetSpec, client: Bfabric, workunit_definition: W
     table = _read_dataset_table(spec)
     check_for_invalid_characters(table=table, invalid_characters=spec.invalid_characters)
 
-    existing = client.reader.query_one("dataset", {"workunitid": registration.workunit_id}, expected_type=Dataset)
+    existing = client.reader.query_one(Dataset, {"workunitid": registration.workunit_id})
     _check_update_existing_policy(
         existing is not None,
         spec.update_existing,
@@ -232,7 +231,7 @@ def _save_link(spec: SaveLinkSpec, client: Bfabric, workunit_definition: Workuni
 def find_default_resource_id(workunit_definition: WorkunitDefinition, client: Bfabric) -> int | None:
     """Finds the default resource's id for the workunit. Maybe in the future, this will be always `None`."""
     workunit_id = workunit_definition.registration.workunit_id  # pyright: ignore[reportOptionalMemberAccess]
-    workunit = client.reader.read_id("workunit", workunit_id, expected_type=Workunit)
+    workunit = client.reader.read_id(Workunit, workunit_id)
     candidate_resources = [
         resource for resource in workunit.resources if resource["name"] not in ["slurm_stdout", "slurm_stderr"]
     ]
@@ -290,7 +289,7 @@ def _get_storage(
     if any(isinstance(spec, CopyResourceSpec) for spec in specs_list):
         if force_storage is None:
             storage_id = workunit_definition.registration.storage_id  # pyright: ignore[reportOptionalMemberAccess]
-            return client.reader.read_id("storage", storage_id, expected_type=Storage)
+            return client.reader.read_id(Storage, storage_id)
         else:
             return Storage(yaml.safe_load(force_storage.read_text()), client=client)
             # TODO replace this later (to avoid versioning issues, i hardcode the above)