+++ tests/bfabric_app_runner/inputs/resolve/test_resolve_bfabric_resource_dataset_specs.py
@@ -85,7 +85,7 @@ def client(mocker):
 
 @pytest.fixture
 def resolver(client):
-    return ResolveBfabricResourceDatasetSpecs(client=client)
+    return ResolveBfabricResourceDatasetSpecs(reader=client.reader)
 
 
 class TestInternals: